requirements.txt
Flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
mysql-connector-python==8.1.0
bcrypt==4.0.1
PyPDF2==3.0.1
python-docx==0.8.11
Werkzeug==2.3.7